requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coolname==2.2.0
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
fastapi==0.112.0
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
kombu==5.4.0
langwatch==0.1.17
nanoid==2.0.0
numpy==2.0.1
openai==1.40.0
pandas==2.2.2
prompt_toolkit==3.0.47
py==1.11.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
retry==0.9.2
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.5
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
zope.event==5.0
zope.interface==7.0.1
